Dynamically allow config's default to get parsed from anything; split out instance methods and class methods
